tri_temp_sync/apps/decision-engine/requirements.txt
# Decision Engine Requirements

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# gRPC
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0

# HTTP Client (para integração com SEM-CSMF e ML-NSMF)
httpx>=0.24.0

# Blockchain (para BC-NSSMF)
web3>=7.0.0
eth-account>=0.13.0
py-solc-x>=1.12.0

# Kafka
kafka-python>=2.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

python-dateutil==2.8.2
requests==2.31.0

